windows compat changes
